+++ src/test/java/com/tomsawyer/perspectives/automation/sikuli/TSAcyclicDesktopPreviewTest.java
@@ -1,10 +1,16 @@
+package com.tomsawyer.perspectives.automation.sikuli;
+
 import java.io.File;
+
 import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.sikuli.basics.Settings;
-
+import com.tomsawyer.perspectives.automation.sikuli.tests.TSDesktopPreviewTestCases;
+import com.tomsawyer.perspectives.automation.sikuli.tests.TSEnvironment;
+import com.tomsawyer.perspectives.automation.sikuli.tests.TSFunctions;
+import com.tomsawyer.perspectives.automation.sikuli.tests.TSTester;
 
 public class TSAcyclicDesktopPreviewTest
 {
