+++ src/test/java/com/tomsawyer/perspectives/automation/sikuli/TSSocialNetworkHtml5PreviewTest.java
@@ -1,9 +1,17 @@
+package com.tomsawyer.perspectives.automation.sikuli;
 import java.io.File;
+
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.AfterClass;
 import org.junit.Test;
 import org.sikuli.basics.Settings;
+
+import com.tomsawyer.perspectives.automation.sikuli.tests.TSAutomationUtils;
+import com.tomsawyer.perspectives.automation.sikuli.tests.TSEnvironment;
+import com.tomsawyer.perspectives.automation.sikuli.tests.TSFunctions;
+import com.tomsawyer.perspectives.automation.sikuli.tests.TSHtml5PreviewTestCases;
+import com.tomsawyer.perspectives.automation.sikuli.tests.TSTester;
 
 
 public class TSSocialNetworkHtml5PreviewTest
